requirements.txt
accelerate
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
comm==0.2.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
debugpy==1.8.7
decorator==5.1.1
distro==1.7.0
Django==4.1.13
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.7.0
drf-spectacular==0.27.2
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
ffmpeg-downloader==0.3.0
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
flatbuffers==24.3.25
fqdn==1.5.1
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.66.1
gunicorn==23.0.0
httplib2==0.20.2
huggingface-hub==0.26.1
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
itypes==1.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keyring==23.5.0
lameenc==1.7.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyxdg==0.27
PyYAML
pyzmq==26.2.0
QtPy==2.4.1
referencing==0.35.1
regex==2024.9.11
reportlab==3.6.8
requests==2.32.3
